refactor(appserver): extract turns + review client
