+++ src/JKToolKit.CodexSDK/AppServer/CodexAppServerClient.cs
@@ -35,12 +35,13 @@ public sealed class CodexAppServerClient : IAsyncDisposable
     private readonly CodexAppServerSkillsAppsClient _skillsAppsClient;
     private readonly CodexAppServerConfigClient _configClient;
     private readonly CodexAppServerFuzzyFileSearchClient _fuzzyFileSearchClient;
+    private readonly CodexAppServerTurnsClient _turnsClient;
+    private readonly CodexAppServerReadOnlyAccessOverridesSupport _readOnlyAccessOverridesSupport = new();
 
     private readonly Channel<AppServerNotification> _globalNotifications;
     private readonly Dictionary<string, CodexTurnHandle> _turnsById = new(StringComparer.Ordinal);
     private int _disposed;
     private int _disconnectSignaled;
-    private int _readOnlyAccessOverridesSupport; // -1 = rejected, 0 = unknown, 1 = supported
     private readonly Task _processExitWatcher;
     private AppServerInitializeResult? _initializeResult;
 
@@ -154,6 +155,13 @@ public sealed class CodexAppServerClient : IAsyncDisposable
         _skillsAppsClient = new CodexAppServerSkillsAppsClient(SendRequestAsync);
         _configClient = new CodexAppServerConfigClient(SendRequestAsync, ExperimentalApiEnabled);
         _fuzzyFileSearchClient = new CodexAppServerFuzzyFileSearchClient(SendRequestAsync, ExperimentalApiEnabled);
+        _turnsClient = new CodexAppServerTurnsClient(
+            options,
+            SendRequestAsync,
+            initializeResult: () => _initializeResult,
+            turnsById: _turnsById,
+            readOnlyAccessOverridesSupport: _readOnlyAccessOverridesSupport,
+            experimentalApiEnabled: ExperimentalApiEnabled);
 
         _globalNotifications = Channel.CreateBounded<AppServerNotification>(new BoundedChannelOptions(options.NotificationBufferCapacity)
         {
@@ -422,79 +430,8 @@ public sealed class CodexAppServerClient : IAsyncDisposable
     /// <summary>
     /// Starts a new turn within the specified thread.
     /// </summary>
-    public async Task<CodexTurnHandle> StartTurnAsync(string threadId, TurnStartOptions options, CancellationToken ct = default)
-    {
-        if (string.IsNullOrWhiteSpace(threadId))
-            throw new ArgumentException("ThreadId cannot be empty or whitespace.", nameof(threadId));
-
-        ArgumentNullException.ThrowIfNull(options);
-
-        ExperimentalApiGuards.ValidateTurnStart(options, experimentalApiEnabled: ExperimentalApiEnabled);
-
-        if (ContainsReadOnlyAccessOverrides(options.SandboxPolicy) &&
-            Volatile.Read(ref _readOnlyAccessOverridesSupport) == -1)
-        {
-            var ua = InitializeResult?.UserAgent ?? "<unknown userAgent>";
-            throw new InvalidOperationException(
-                $"turn/start sandboxPolicy ReadOnlyAccess overrides were previously rejected by this app-server build. userAgent='{ua}'. Do not send ReadOnlyAccess fields unless your Codex app-server supports them.");
-        }
-
-        var turnStartParams = new TurnStartParams
-        {
-            ThreadId = threadId,
-            Input = options.Input.Select(i => i.Wire).ToArray(),
-            Cwd = options.Cwd,
-            ApprovalPolicy = options.ApprovalPolicy?.Value,
-            SandboxPolicy = options.SandboxPolicy,
-            Model = options.Model?.Value,
-            Effort = options.Effort?.Value,
-            Summary = options.Summary,
-            Personality = options.Personality,
-            OutputSchema = options.OutputSchema,
-            CollaborationMode = options.CollaborationMode
-        };
-
-        JsonElement result;
-        try
-        {
-            result = await SendRequestAsync("turn/start", turnStartParams, ct);
-        }
-        catch (JsonRpcRemoteException ex) when (ex.Error.Code == -32602 && ContainsReadOnlyAccessOverrides(options.SandboxPolicy))
-        {
-            Interlocked.Exchange(ref _readOnlyAccessOverridesSupport, -1);
-            var ua = InitializeResult?.UserAgent ?? "<unknown userAgent>";
-            var sandboxJson = JsonSerializer.Serialize(options.SandboxPolicy, CreateDefaultSerializerOptions());
-            var data = ex.Error.Data is { ValueKind: not JsonValueKind.Null and not JsonValueKind.Undefined }
-                ? $" Data: {ex.Error.Data.Value.GetRawText()}"
-                : string.Empty;
-
-            throw new InvalidOperationException(
-                $"turn/start rejected sandboxPolicy parameters (likely unsupported by this Codex app-server build). userAgent='{ua}'. sandboxPolicy={sandboxJson}. Error: {ex.Error.Code}: {ex.Error.Message}.{data}",
-                ex);
-        }
-
-        if (ContainsReadOnlyAccessOverrides(options.SandboxPolicy))
-        {
-            Interlocked.Exchange(ref _readOnlyAccessOverridesSupport, 1);
-        }
-
-        var turnId = ExtractTurnId(result);
-        if (string.IsNullOrWhiteSpace(turnId))
-        {
-            throw new InvalidOperationException(
-                $"turn/start returned no turn id. Raw result: {result}");
-        }
-
-        return CreateTurnHandle(threadId, turnId);
-    }
-
-    private static bool ContainsReadOnlyAccessOverrides(SandboxPolicy? policy) =>
-        policy switch
-        {
-            SandboxPolicy.ReadOnly r => r.Access is not null,
-            SandboxPolicy.WorkspaceWrite w => w.ReadOnlyAccess is not null,
-            _ => false
-        };
+    public Task<CodexTurnHandle> StartTurnAsync(string threadId, TurnStartOptions options, CancellationToken ct = default) =>
+        _turnsClient.StartTurnAsync(threadId, options, ct);
 
     /// <summary>
     /// Steers an in-progress turn by appending input items.
@@ -503,11 +440,8 @@ public sealed class CodexAppServerClient : IAsyncDisposable
     /// Steering is best-effort and may race with turn completion. Cancellation stops waiting for the response but does
     /// not guarantee the server did not apply the steer request.
     /// </remarks>
-    public async Task<string> SteerTurnAsync(TurnSteerOptions options, CancellationToken ct = default)
-    {
-        var result = await SteerTurnRawAsync(options, ct);
-        return result.TurnId;
-    }
+    public Task<string> SteerTurnAsync(TurnSteerOptions options, CancellationToken ct = default) =>
+        _turnsClient.SteerTurnAsync(options, ct);
 
     /// <summary>
     /// Steers an in-progress turn by appending input items and returns the raw JSON result payload.
@@ -516,89 +450,14 @@ public sealed class CodexAppServerClient : IAsyncDisposable
     /// Steering is best-effort and may race with turn completion. Cancellation stops waiting for the response but does
     /// not guarantee the server did not apply the steer request.
     /// </remarks>
-    public async Task<TurnSteerResult> SteerTurnRawAsync(TurnSteerOptions options, CancellationToken ct = default)
-    {
-        ArgumentNullException.ThrowIfNull(options);
-        if (string.IsNullOrWhiteSpace(options.ThreadId))
-            throw new ArgumentException("ThreadId cannot be empty or whitespace.", nameof(options.ThreadId));
-        if (string.IsNullOrWhiteSpace(options.ExpectedTurnId))
-            throw new ArgumentException("ExpectedTurnId cannot be empty or whitespace.", nameof(options.ExpectedTurnId));
-
-        try
-        {
-            var raw = await SendRequestAsync(
-                "turn/steer",
-                BuildTurnSteerParams(options),
-                ct);
-
-            return new TurnSteerResult
-            {
-                TurnId = ExtractTurnId(raw) ?? options.ExpectedTurnId,
-                Raw = raw
-            };
-        }
-        catch (JsonRpcRemoteException ex)
-        {
-            var ua = InitializeResult?.UserAgent;
-            throw new CodexAppServerRequestFailedException(
-                method: "turn/steer",
-                errorCode: ex.Error.Code,
-                errorMessage: $"{ex.Error.Message} (expectedTurnId='{options.ExpectedTurnId}')",
-                errorData: ex.Error.Data,
-                userAgent: ua,
-                innerException: ex);
-        }
-    }
+    public Task<TurnSteerResult> SteerTurnRawAsync(TurnSteerOptions options, CancellationToken ct = default) =>
+        _turnsClient.SteerTurnRawAsync(options, ct);
 
     /// <summary>
     /// Starts a review via the app-server.
     /// </summary>
-    public async Task<ReviewStartResult> StartReviewAsync(ReviewStartOptions options, CancellationToken ct = default)
-    {
-        ArgumentNullException.ThrowIfNull(options);
-        if (string.IsNullOrWhiteSpace(options.ThreadId))
-            throw new ArgumentException("ThreadId cannot be empty or whitespace.", nameof(options.ThreadId));
-        ArgumentNullException.ThrowIfNull(options.Target);
-
-        JsonElement result;
-        try
-        {
-            result = await SendRequestAsync(
-                "review/start",
-                BuildReviewStartParams(options),
-                ct);
-        }
-        catch (JsonRpcRemoteException ex)
-        {
-            var ua = InitializeResult?.UserAgent;
-            throw new CodexAppServerRequestFailedException(
-                method: "review/start",
-                errorCode: ex.Error.Code,
-                errorMessage: ex.Error.Message,
-                errorData: ex.Error.Data,
-                userAgent: ua,
-                innerException: ex);
-        }
-
-        var reviewThreadId = GetStringOrNull(result, "reviewThreadId");
-
-        var turnObj = TryGetObject(result, "turn") ?? result;
-        var turnId = ExtractTurnId(turnObj);
-        if (string.IsNullOrWhiteSpace(turnId))
-        {
-            throw new InvalidOperationException(
-                $"review/start returned no turn id. Raw result: {result}");
-        }
-
-        var turnThreadId = ExtractThreadId(turnObj) ?? reviewThreadId ?? options.ThreadId;
-
-        return new ReviewStartResult
-        {
-            Turn = CreateTurnHandle(turnThreadId, turnId),
-            ReviewThreadId = reviewThreadId,
-            Raw = result
-        };
-    }
+    public Task<ReviewStartResult> StartReviewAsync(ReviewStartOptions options, CancellationToken ct = default) =>
+        _turnsClient.StartReviewAsync(options, ct);
 
     /// <summary>
     /// Starts a review via the app-server.
@@ -607,7 +466,7 @@ public sealed class CodexAppServerClient : IAsyncDisposable
     /// This is an alias for <see cref="StartReviewAsync"/> to better align with exec-mode naming (<c>ReviewAsync</c>).
     /// </remarks>
     public Task<ReviewStartResult> ReviewAsync(ReviewStartOptions options, CancellationToken ct = default) =>
-        StartReviewAsync(options, ct);
+        _turnsClient.StartReviewAsync(options, ct);
 
     internal static TurnSteerParams BuildTurnSteerParams(TurnSteerOptions options) =>
         new()
@@ -629,37 +488,6 @@ public sealed class CodexAppServerClient : IAsyncDisposable
                 _ => null
             }
         };
-
-    private CodexTurnHandle CreateTurnHandle(string threadId, string turnId)
-    {
-        var handle = new CodexTurnHandle(
-            threadId,
-            turnId,
-            interrupt: c => InterruptAsync(threadId, turnId, c),
-            steer: (input, c) => SteerTurnAsync(new TurnSteerOptions { ThreadId = threadId, ExpectedTurnId = turnId, Input = input }, c),
-            steerRaw: (input, c) => SteerTurnRawAsync(new TurnSteerOptions { ThreadId = threadId, ExpectedTurnId = turnId, Input = input }, c),
-            onDispose: () =>
-            {
-                lock (_turnsById)
-                {
-                    _turnsById.Remove(turnId);
-                }
-            },
-            bufferCapacity: _options.NotificationBufferCapacity);
-
-        lock (_turnsById)
-        {
-            _turnsById[turnId] = handle;
-        }
-
-        return handle;
-    }
-
-    private Task InterruptAsync(string threadId, string turnId, CancellationToken ct) =>
-        SendRequestAsync(
-            "turn/interrupt",
-            new TurnInterruptParams { ThreadId = threadId, TurnId = turnId },
-            ct);
 
     private ValueTask OnRpcNotificationAsync(JsonRpcNotification notification)
     {
