+++ src/JKToolKit.CodexSDK/AppServer/Internal/CodexAppServerReadOnlyAccessOverridesSupport.cs
@@ -1,0 +1,8 @@
+namespace JKToolKit.CodexSDK.AppServer.Internal;
+
+internal sealed class CodexAppServerReadOnlyAccessOverridesSupport
+{
+    // -1 = rejected, 0 = unknown, 1 = supported
+    public int Value;
+}
+
